Finally, everthing is completed. Files updated: 1. ShowSettngs.java 2. showsettngs.xml
A new button "Save Server Url Without Logging In" is shown on this screen so that the user can access the Depot server as general public.

3. CatalogList.java
4. CatalogEntry.java
5. Product.java
6. NewProduct.java
7. Catalog.java
8. CatalogByTile.java

These files are changed mainly to allow the user to operate under both online and offline modes. Also, all four CRUD operations are supported under both modes.